+++ pages/index.tsx
@@ -9,13 +9,14 @@ import MainList from '../components/main/MainList';
 import MainTags from '../components/main/MainTags';
 import RankList from '../components/main/RankList';
 import Notice from '../components/rightLayout/Notice';
+import { TitleMain } from '../components/common/TitleMain';
 
 const Home: NextPage = () => {
    return (
       <>
          <Category />
          <CenterSection>
-            <Title />
+            <TitleMain />
             <MainTags />
             <MainList />
          </CenterSection>
